backend/queueMgt/requirements.txt
python-dotenv
Flask
google-cloud-firestore
google-auth
fastapi
pydantic
firebase-admin